% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMikolinski.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Example Dataset}
\format{
A data frame with columns:
\describe{
\item{StateProvince}{State}
\item{decimalLongitude}{Longitude of observation}
\item{decimalLatitude}{Latitude of observation}

}
}
\source{
Simulated data for illustrative purposes.
}
\usage{
dataset
}
\description{
This dataset provides an example of data with ID, Name, Age, and Gender columns.
}
\examples{
data(dataset)
# Viewing the structure of the dataset
str(dataset)

# Summarizing the dataset
summary(dataset)

# Accessing specific variables
dataset$decimalLatitude
dataset$decimalLongitude
dataset$stateProvince
}
\keyword{datasets}
