% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMikolinski.R
\name{greater_mlatlongdf}
\alias{greater_mlatlongdf}
\title{Above positive slope}
\usage{
greater_mlatlongdf(a, b, c, d, e, f)
}
\arguments{
\item{a}{Data frame}

\item{b}{Latitude 1 (smaller)}

\item{c}{Longitude 1 (smaller)}

\item{d}{Latitude 2 (larger)}

\item{e}{Longitude 2 (larger)}

\item{f}{State}
}
\value{
Subset of data frame that is above the line
}
\description{
This function gives all data points in a data frame that are above the line created by the 2 sets of coordinates
}
\examples{
greater_mlatlongdf(dataset, 33.288275, -82.899328, 35.197685, -82.226109, "South Carolina")
}
