% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMikolinski.R
\name{nless_mlatlongdf}
\alias{nless_mlatlongdf}
\title{Below negative slope}
\usage{
nless_mlatlongdf(a, b, c, d, e, f)
}
\arguments{
\item{a}{Data frame}

\item{b}{Latitude 1 (larger)}

\item{c}{Longitude 1 (larger)}

\item{d}{Latitude 2 (smaller)}

\item{e}{Longitude 2 (smaller)}

\item{f}{State}
}
\value{
Subset of data frame that is below the line
}
\description{
This function gives all data points in a data frame that are below the line created by the 2 sets of coordinates
}
\examples{
nless_mlatlongdf(dataset, 35.197685, -82.226109, 33.288275, -82.899328, "South Carolina")
}
